Working on persistant notifications
Due to a Heroku "feature", it has become necessary to store
notificaitons in a DB. Basically, Heroku restarts the service
periodically and when server resources shift around. The solution
is to save notifications in a (possibly) separate DB when ice
exits and reload them into memory when it starts. To do this,
I'm subclassing Flask.